+++ grid2op/Space/RandomObject.py
@@ -24,6 +24,23 @@ class RandomObject(object):
     seed_used: ``int``
         The seed used throughout the episode in case of non deterministic observations or action.
 
+    Notes
+    -----
+        In order to be reproducible, and to make proper use of the
+        :func:`BaseAgent.seed` capabilities, you must absolutely NOT use the `random` python module (which will not
+        be seeded) nor the `np.random` module and avoid any other "sources" of pseudo random numbers.
+
+        You can adapt your code the following way. Instead of using `np.random` use `self.space_prng`.
+
+        For example, if you wanted to write
+        `np.random.randint(1,5)` replace it by `self.space_prng.randint(1,5)`. It is the same for `np.random.normal()`
+        that is
+        replaced by `self.space_prng.normal()`.
+
+        You have an example of such usage in :func:`RandomAgent.my_act`.
+
+        If you really need other sources of randomness (for example if you use tensorflow or torch) we strongly
+        recommend you to overload the :func:`BaseAgent.seed` accordingly. In that
     """
     def __init__(self):
         self.space_prng = np.random.RandomState()
@@ -31,13 +48,22 @@ class RandomObject(object):
 
     def seed(self, seed):
         """
-        Use to set the seed in case of non deterministic observations.
-        :param seed:
-        :return:
+        Set the seed of the source of pseudo random number used for this RandomObject.
+
+        Parameters
+        ----------
+        seed: ``int``
+            The seed to be set.
+
+        Returns
+        -------
+        res: ``tuple``
+            The associated tuple of seeds used. Tuples are returned because in some cases, multiple objects are seeded
+            with the same call to :func:`RandomObject.seed`
+
         """
         self.seed_used = seed
         if self.seed_used is not None:
             # in this case i have specific seed set. So i force the seed to be deterministic.
             self.space_prng.seed(seed=self.seed_used)
-        return self.seed_used
-
+        return self.seed_used,